Added GyroReset
This SHOULD be working code for the entire robot. I used forward declarations which is not needed in Java but it is used in C so that I am able to use functions. The main task is as small as possible because that code is copied line per line into the memory of the Cortex and to save memory, make the main() task as small as possible. The functions are only used once they are called saving the memory. I was reluctant to use public/global variables because they van be changed anywhere and they are VERY BAD programming practice. Avoid them at all costs. I ended up using them because otherwise I would either initialize the same variable every time the Cortex loops the code, hogging memory or I would have to use a static variable which requires me to make  classes which are a lot more work. I have "_speed" because that is private variable which can only be accessed within the function and its good practice to add the underscore. All variables are written with the first word lower case and all thereafter with the first letter capitalized like "gyroInit". I think that's it. If you got any questions feel free to comment.